requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.2
django-vite==3.0.3
idna==3.6
inertia-django==0.6.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
